fixed all clang warnings from -Weverything

+++ util/util.c
@@ -5,7 +5,7 @@
 #include "util.h"
 
 char *make_string(char *str) {
-    int len = strlen(str);
+    unsigned long len = strlen(str);
     char *buf = malloc((1 + len) * sizeof(char));
     memcpy(buf, str, len);
     return buf;
